requirements.txt
tweepy
schedule
requests
python-dotenv